test: move source-related tests to datasource.rs
